requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.5
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.165.1
cryptography==45.0.6
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
Django==5.2.5
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.10.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.1
json_repair==0.25.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
litellm==1.74.9
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.99.9
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
overrides==7.7.0
packaging==25.0
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
portalocker==2.7.0
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyvis==0.3.2
pywin32==311
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.13
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
